Implementing npc generation

+++ backend/src/main/java/com/github/acolote1998/humble_gladiators_2/characters/model/Inventory.java
@@ -48,7 +48,7 @@ public class Inventory {
     @JoinColumn(name = "inventory_id")
     List<WeaponInstance> weapons;
 
-    Long gold;
+    Integer gold;
 
     @CreationTimestamp
     LocalDateTime createdAt; // Auto-managed by JPA
